+++ Example/BSImagePickerExample/ViewController.swift
@@ -39,7 +39,77 @@ class ViewController: UIViewController {
                 print("Cancel: \(assets)")
             }, finish: { (assets: [PHAsset]) -> Void in
                 print("Finish: \(assets)")
-        }, completion: nil)
+            }, completion: nil)
+    }
+    
+    @IBAction func showCustomDataSourcePicker(sender: UIButton) {
+        // Find some PHAssets to use. In the real world these would probably come from the camera
+        // Or some other source. If you want to use your own fetch results for the picker, there is an initalizer for that as well
+        // You can also supply an array of assets that should be selected on presentation
+        let fetchOptions = PHFetchOptions()
+        
+        // Camera roll collection
+        let cameraRollCollection = PHAssetCollection.fetchAssetCollectionsWithType(.SmartAlbum, subtype: .SmartAlbumUserLibrary, options: fetchOptions).objectAtIndex(0) as! PHAssetCollection
+        
+        // Pick out 3 PHAssets
+        let assets = PHAsset.fetchAssetsInAssetCollection(cameraRollCollection, options: fetchOptions)
+        let first = assets.objectAtIndex(0) as! PHAsset
+        let second = assets.objectAtIndex(1) as! PHAsset
+        let third = assets.objectAtIndex(2) as! PHAsset
+        
+        // Create transient asset collection
+        let selectedAssets = [first, second, third]
+        let transientCollection = PHAssetCollection.transientAssetCollectionWithAssets(selectedAssets, title: "Custom assets")
+        
+        let vc = BSImagePickerViewController(assetCollection: transientCollection, selections: selectedAssets)
+        
+        bs_presentImagePickerController(vc, animated: true,
+            select: { (asset: PHAsset) -> Void in
+                println("Selected: \(asset)")
+            }, deselect: { (asset: PHAsset) -> Void in
+                println("Deselected: \(asset)")
+            }, cancel: { (assets: [PHAsset]) -> Void in
+                println("Cancel: \(assets)")
+            }, finish: { (assets: [PHAsset]) -> Void in
+                println("Finish: \(assets)")
+            }, completion: nil)
+    }
+    
+    @IBAction func showCustomImagePicker(sender: UIButton) {
+        let vc = BSImagePickerViewController()
+        vc.maxNumberOfSelections = 6
+        
+        vc.albumButton.tintColor = UIColor.greenColor()
+        vc.cancelButton.tintColor = UIColor.redColor()
+        vc.doneButton.tintColor = UIColor.purpleColor()
+        vc.selectionCharacter = "✓"
+        vc.selectionFillColor = UIColor.blackColor()
+        vc.selectionStrokeColor = UIColor.yellowColor()
+        vc.selectionShadowColor = UIColor.redColor()
+        vc.selectionTextAttributes[NSForegroundColorAttributeName] = UIColor.lightGrayColor()
+        vc.cellsPerRow = {(verticalSize: UIUserInterfaceSizeClass, horizontalSize: UIUserInterfaceSizeClass) -> Int in
+            switch (verticalSize, horizontalSize) {
+            case (.Compact, .Regular): // iPhone5-6 portrait
+                return 2
+            case (.Compact, .Compact): // iPhone5-6 landscape
+                return 2
+            case (.Regular, .Regular): // iPad portrait/landscape
+                return 3
+            default:
+                return 2
+            }
+        }
+        
+        bs_presentImagePickerController(vc, animated: true,
+            select: { (asset: PHAsset) -> Void in
+                println("Selected: \(asset)")
+            }, deselect: { (asset: PHAsset) -> Void in
+                println("Deselected: \(asset)")
+            }, cancel: { (assets: [PHAsset]) -> Void in
+                println("Cancel: \(assets)")
+            }, finish: { (assets: [PHAsset]) -> Void in
+                println("Finish: \(assets)")
+            }, completion: nil)
     }
 }