+++ Pod/Classes/Controller/BSImagePickerViewController.swift
@@ -23,68 +23,265 @@
 import UIKit
 import Photos
 
-public class BSImagePickerViewController : UINavigationController, BSImagePickerSettings {
-    internal lazy var photosViewController: PhotosViewController = {
-        // Get path for BSImagePicker bundle
-        let bundlePath = NSBundle(forClass: PhotosViewController.self).pathForResource("BSImagePicker", ofType: "bundle")
-        let bundle: NSBundle?
-        
-        // Load bundle
-        if let bundlePath = bundlePath {
-            bundle = NSBundle(path: bundlePath)
+/**
+BSImagePickerViewController.
+Use settings or buttons to customize it to your needs.
+*/
+public final class BSImagePickerViewController : UINavigationController, BSImagePickerSettings {
+    private let settings = Settings()
+    
+    private var doneBarButton: UIBarButtonItem = UIBarButtonItem(barButtonSystemItem: .Done, target: nil, action: nil)
+    private var cancelBarButton: UIBarButtonItem = UIBarButtonItem(barButtonSystemItem: .Cancel, target: nil, action: nil)
+    private let albumTitleView: AlbumTitleView = bundle.loadNibNamed("AlbumTitleView", owner: nil, options: nil).first as! AlbumTitleView
+    private var dataSource: SelectableDataSource?
+    private let selections: [PHAsset]
+    
+    static let bundle: NSBundle = NSBundle(path: NSBundle(forClass: PhotosViewController.self).pathForResource("BSImagePicker", ofType: "bundle")!)!
+    
+    lazy var photosViewController: PhotosViewController = {
+        let dataSource: SelectableDataSource
+        if self.dataSource != nil {
+            dataSource = self.dataSource!
         } else {
-            bundle = nil
+            dataSource = BSImagePickerViewController.defaultDataSource()
         }
         
-        let storyboard = UIStoryboard(name: "Photos", bundle: bundle)
+        let vc = PhotosViewController(dataSource: dataSource, settings: self.settings, selections: self.selections)
         
-        return storyboard.instantiateInitialViewController() as! PhotosViewController
+        vc.doneBarButton = self.doneBarButton
+        vc.cancelBarButton = self.cancelBarButton
+        vc.albumTitleView = self.albumTitleView
+        
+        return vc
     }()
     
-    public convenience init() {
-        self.init(nibName: nil, bundle: nil)
-    }
-
-    required public init(coder aDecoder: NSCoder) {
-        super.init(coder: aDecoder)
-    }
-    
-    override init(nibName nibNameOrNil: String?, bundle nibBundleOrNil: NSBundle?) {
-        super.init(nibName: nibNameOrNil, bundle: nibBundleOrNil)
-    }
-    
-    public override func loadView() {
-        super.loadView()
-        
-        view.backgroundColor = UIColor.whiteColor()
-    }
-    
-    public override func viewWillAppear(animated: Bool) {
-        super.viewWillAppear(animated)
-        
-        updateViewControllerToStatus(PHPhotoLibrary.authorizationStatus())
-    }
-    
-    func updateViewControllerToStatus(status: PHAuthorizationStatus) {
+    class func authorize(status: PHAuthorizationStatus = PHPhotoLibrary.authorizationStatus(), fromViewController: UIViewController, completion: () -> Void) {
         switch status {
         case .Authorized:
-            // We are authorized. Push photos view controller
-            viewControllers = []
-            pushViewController(self.photosViewController, animated: false)
-            
+            // We are authorized. Run block
+            completion()
         case .NotDetermined:
             // Ask user for permission
             PHPhotoLibrary.requestAuthorization({ (status) -> Void in
                 dispatch_async(dispatch_get_main_queue(), { () -> Void in
-                    self.updateViewControllerToStatus(status)
+                    self.authorize(status: status, fromViewController: fromViewController, completion: completion)
                 })
             })
         default: ()
+            dispatch_async(dispatch_get_main_queue(), { () -> Void in
+                // Set up alert controller with some default strings. These should probably be overriden in application localizeable strings.
+                // If you don't enjoy my Swenglish that is ^^
+                let alertController = UIAlertController(title: NSLocalizedString("imagePickerNoCameraAccessTitle", value: "Can't access Photos", comment: "Alert view title"),
+                    message: NSLocalizedString("imagePickerNoCameraAccessMessage", value: "You need to enable Photos access in application settings.", comment: "Alert view message"),
+                    preferredStyle: .Alert)
+                
+                let cancelAction = UIAlertAction(title: NSLocalizedString("imagePickerNoCameraAccessCancelButton", value: "Cancel", comment: "Cancel button title"), style: .Cancel, handler:nil)
+                
+                let settingsAction = UIAlertAction(title: NSLocalizedString("imagePickerNoCameraAccessSettingsButton", value: "Settings", comment: "Settings button title"), style: .Default, handler: { (action) -> Void in
+                    let url = NSURL(string: UIApplicationOpenSettingsURLString)
+                    if let url = url where UIApplication.sharedApplication().canOpenURL(url) {
+                        UIApplication.sharedApplication().openURL(url)
+                    }
+                })
+                
+                alertController.addAction(cancelAction)
+                alertController.addAction(settingsAction)
+                
+                fromViewController.presentViewController(alertController, animated: true, completion: nil)
+            })
         }
     }
     
+    /**
+    Want it to show your own custom fetch results? Make sure the fetch results are of PHAssetCollections
+    :param: fetchResults PHFetchResult of PHAssetCollections
+    */
+    public convenience init(fetchResults: [PHFetchResult]) {
+        self.init(dataSource: FetchResultsDataSource(fetchResults: fetchResults))
+    }
+    
+    /**
+    Do you have an asset collection you want to select from? Use this initializer!
+    :param: assetCollection The PHAssetCollection you want to select from
+    :param: selections Selected assets
+    */
+    public convenience init(assetCollection: PHAssetCollection, selections: [PHAsset] = []) {
+        self.init(dataSource: AssetCollectionDataSource(assetCollection: assetCollection), selections: selections)
+    }
+    
+    /**
+    Sets up an classic image picker with results from camera roll and albums
+    */
+    public convenience init() {
+        self.init(dataSource: nil)
+    }
+    
+    /**
+    You should probably use one of the convenience inits
+    :param: dataSource The data source for the albums
+    :param: selections Any PHAsset you want to seed the picker with as selected
+    */
+    public required init(dataSource: SelectableDataSource?, selections: [PHAsset] = []) {
+        if let dataSource = dataSource {
+            self.dataSource = dataSource
+        }
+        
+        self.selections = selections
+        
+        super.init(nibName: nil, bundle: nil)
+    }
+
+    /**
+    https://www.youtube.com/watch?v=dQw4w9WgXcQ
+    */
+    required public init(coder aDecoder: NSCoder) {
+        dataSource = BSImagePickerViewController.defaultDataSource()
+        selections = []
+        super.init(coder: aDecoder)
+    }
+    
+    /**
+    Load view. See apple documentation
+    */
+    public override func loadView() {
+        super.loadView()
+        
+        // TODO: Settings
+        view.backgroundColor = UIColor.whiteColor()
+        
+        // Make sure we really are authorized
+        if PHPhotoLibrary.authorizationStatus() == .Authorized {
+            setViewControllers([photosViewController], animated: false)
+        }
+    }
+    
+    private static func defaultDataSource() -> SelectableDataSource {
+        let fetchOptions = PHFetchOptions()
+        
+        // Camera roll fetch result
+        let cameraRollResult = PHAssetCollection.fetchAssetCollectionsWithType(.SmartAlbum, subtype: .SmartAlbumUserLibrary, options: fetchOptions)
+        
+        // Albums fetch result
+        let albumResult = PHAssetCollection.fetchAssetCollectionsWithType(.Album, subtype: .Any, options: fetchOptions)
+        
+        return FetchResultsDataSource(fetchResults: [cameraRollResult, albumResult])
+    }
+    
     // MARK: ImagePickerSettings proxy
-    public var selectionClosure: ((asset: PHAsset) -> Void)? {
+    /**
+    See BSImagePicketSettings for documentation
+    */
+    public var maxNumberOfSelections: Int {
+        get {
+            return settings.maxNumberOfSelections
+        }
+        set {
+            settings.maxNumberOfSelections = newValue
+        }
+    }
+    
+    /**
+    See BSImagePicketSettings for documentation
+    */
+    public var selectionCharacter: Character? {
+        get {
+            return settings.selectionCharacter
+        }
+        set {
+            settings.selectionCharacter = newValue
+        }
+    }
+    
+    /**
+    See BSImagePicketSettings for documentation
+    */
+    public var selectionFillColor: UIColor {
+        get {
+            return settings.selectionFillColor
+        }
+        set {
+            settings.selectionFillColor = newValue
+        }
+    }
+    
+    /**
+    See BSImagePicketSettings for documentation
+    */
+    public var selectionStrokeColor: UIColor {
+        get {
+            return settings.selectionStrokeColor
+        }
+        set {
+            settings.selectionStrokeColor = newValue
+        }
+    }
+    
+    /**
+    See BSImagePicketSettings for documentation
+    */
+    public var selectionShadowColor: UIColor {
+        get {
+            return settings.selectionShadowColor
+        }
+        set {
+            settings.selectionShadowColor = newValue
+        }
+    }
+    
+    /**
+    See BSImagePicketSettings for documentation
+    */
+    public var selectionTextAttributes: [NSObject: AnyObject] {
+        get {
+            return settings.selectionTextAttributes
+        }
+        set {
+            settings.selectionTextAttributes = newValue
+        }
+    }
+    
+    /**
+    See BSImagePicketSettings for documentation
+    */
+    public var cellsPerRow: (verticalSize: UIUserInterfaceSizeClass, horizontalSize: UIUserInterfaceSizeClass) -> Int {
+        get {
+            return settings.cellsPerRow
+        }
+        set {
+            settings.cellsPerRow = newValue
+        }
+    }
+    
+    // MARK: Buttons
+    /**
+    Cancel button
+    */
+    public var cancelButton: UIBarButtonItem {
+        get {
+            return self.cancelBarButton
+        }
+    }
+    
+    /**
+    Done button
+    */
+    public var doneButton: UIBarButtonItem {
+        get {
+            return self.doneBarButton
+        }
+    }
+    
+    /**
+    Album button
+    */
+    public var albumButton: UIButton {
+        get {
+            return self.albumTitleView.albumButton
+        }
+    }
+    
+    // MARK: Closures
+    var selectionClosure: ((asset: PHAsset) -> Void)? {
         get {
             return photosViewController.selectionClosure
         }
@@ -92,7 +289,7 @@ public class BSImagePickerViewController : UINavigationController, BSImagePicker
             photosViewController.selectionClosure = newValue
         }
     }
-    public var deselectionClosure: ((asset: PHAsset) -> Void)? {
+    var deselectionClosure: ((asset: PHAsset) -> Void)? {
         get {
             return photosViewController.deselectionClosure
         }
@@ -100,7 +297,7 @@ public class BSImagePickerViewController : UINavigationController, BSImagePicker
             photosViewController.deselectionClosure = newValue
         }
     }
-    public var cancelClosure: ((assets: [PHAsset]) -> Void)? {
+    var cancelClosure: ((assets: [PHAsset]) -> Void)? {
         get {
             return photosViewController.cancelClosure
         }
@@ -108,21 +305,12 @@ public class BSImagePickerViewController : UINavigationController, BSImagePicker
             photosViewController.cancelClosure = newValue
         }
     }
-    public var finishClosure: ((assets: [PHAsset]) -> Void)? {
+    var finishClosure: ((assets: [PHAsset]) -> Void)? {
         get {
             return photosViewController.finishClosure
         }
         set {
             photosViewController.finishClosure = newValue
-        }
-    }
-    
-    public var maxNumberOfSelections: Int {
-        get {
-            return photosViewController.maxNumberOfSelections
-        }
-        set {
-            photosViewController.maxNumberOfSelections = newValue
         }
     }
 }