+++ CHANGELOG.md
@@ -1,3 +1,20 @@
+# 1.4.1
+* Fixed crash when presenting for the first time
+
+# 1.4.0
+* Allows you to initialize the picker with your own fetch results or asset collection. And an array of assets that should be selected on presentation.
+
+# 1.3.0
+* Setting for cells per row
+* Performance tweaks (final classes)
+
+# 1.2.0
+* More settings for you to tweak (relating to selection)
+
+# 1.1.0
+* Exposed cancel, album and done buttons for you to customize
+* selectionCharacter property that you can set if you don’t want the numbered selection
+
 # 1.0.1
 * Fixed a crash with an IUO in the Photos framework